requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
SQLAlchemy==2.0.23
weaviate-client==4.4.0
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Shopping Agent Dependencies
adk-python>=0.1.0
mcp>=0.1.0
openai>=1.0.0
anthropic>=0.1.0
google-generativeai>=0.3.0
websockets>=11.0
aiohttp>=3.9.0